proxy/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
pydantic==2.9.0
pydantic-settings==2.5.2
PyJWT==2.9.0
jinja2==3.1.4
aiosqlite==0.20.0
python-multipart==0.0.12
